feat(planning): encode four more programmes, and make min_credits fractional
Entrepreneurship and Management, Event Management, Executive Coaching and
Organizational Consulting, and Marketing and Strategic Communications take the
encoded set from three to seven. All four reconcile against their stated totals.

`Requirement.min_credits` had to become fractional alongside `Course.credits`:
Event Management's core curriculum is seven courses totalling 16.5, which an
integer column rounds into a requirement the bulletin does not state.

None of the four needed a new rule. Three shapes recurred and all were already
covered — a closed elective pool counted in credits, a two-option capstone as
`one_track` with one course per track, and a wholly required curriculum as
`all_of`. Executive Coaching's "Module 1 / Module 2" headings order a cohort's
residencies rather than offering a choice, so they are presentation and not
encoded as anything.

**The programme worksheet was being read through a filter that quietly under-read
it.** The extraction asked for three heading prefixes — Program Requirements,
Concentrations, Areas of Study — and Event Management publishes its elective
courses under `Tracks`. On that evidence I had concluded its 15 elective credits
were unenumerable and that the programme could not be faithfully encoded, which
was wrong: the courses are on the page, in full, one heading over. Re-querying by
exclusion instead of by guessed prefixes turned up a fourth name as well,
`Specializations`. Same failure as the four-digit course code — a pattern fitted
to what had already been seen, failing silently on what had not.

Global Affairs is left unencoded, and this one is a real gap rather than a
misread. Each of its six concentrations is one named required course plus five
chosen from a pool of thirty-odd. `TrackRule` can say "choose N from these" or
"take all of these", not "this one, and N of those" — and encoding it as the
former would let a student satisfy a concentration while skipping the course it
requires. A wrong pass in a degree audit is the failure this engine exists to
prevent, so it waits for the rule rather than getting an approximation.

Verified: pytest 361, authz_probe 28/28, mission_probe 36/36, and all seven
programmes write and reconcile.

+++ api/app/models/academic.py
@@ -141,7 +141,10 @@ class Requirement(Base, TimestampMixin):
     kind: Mapped[RequirementKind] = mapped_column(
         SAEnum(RequirementKind, name="requirement_kind"), nullable=False
     )
-    min_credits: Mapped[int] = mapped_column(Integer, nullable=False)
+    # Float for the same reason Course.credits is: a group of 1.5-credit courses sums to a
+    # half. Event Management's core curriculum is 16.5, which an integer column rounds into a
+    # requirement the bulletin does not state.
+    min_credits: Mapped[float] = mapped_column(Float, nullable=False)
     min_courses: Mapped[int | None] = mapped_column(Integer)
     sort_order: Mapped[int] = mapped_column(Integer, default=0, nullable=False)
     notes: Mapped[str | None] = mapped_column(Text)
